backend/pyproject.toml
[project]
name = "photo-triage-backend"
version = "0.1.0"
description = "Local-first photo analysis backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "structlog>=24.1.0",
    "pillow>=10.2.0",
    "imagehash>=4.3.1",
    "opencv-python-headless>=4.9.0",
    "llama-cpp-python>=0.2.50",
    "send2trash>=1.8.0",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "huggingface-hub>=1.3.1",
    "psutil>=7.2.1",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
